requirements.txt
Django==4.1.7
dj-rest-auth==3.0.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==4.8.0
six==1.16.0
django-allauth==0.52.0
python-decouple==3.8
psycopg2-binary==2.9.5
gunicorn==20.1.0
django_heroku==0.3.1
dj_database_url==1.2.0
whitenoise==6.4.0
openai==0.27.2